Move binance-specific test to test_binance.py
